# perf(jsx-runtime): optimize `createVNode#defaultProps`